+++ ios/Frameworks/KMPWorkManager.xcframework/ios-arm64/KMPWorkManager.framework/Headers/KMPWorkManager.h
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6b18bb7d2f73411bcdc6e2ae1f43a8053adb4df3d2eaff3a95d25881721494e7
-size 416791
+oid sha256:a413e3750b6e1f96f600155705f9aadf7e84ccf652fc8cc25d0ed2de78b11e18
+size 297371